# openapi/depotflux-v1.json
{
  "components": {
    "schemas": {
      "AgentBackend": {
        "enum": [
          "rule"
        ],
        "title": "AgentBackend",
        "type": "string"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptimizationMode": {
        "enum": [
          "selfish",
          "altruistic"
        ],
        "title": "OptimizationMode",
        "type": "string"
      },
...
      "RunResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_backend": {
            "$ref": "#/components/schemas/AgentBackend"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_reference": {
            "title": "Input Reference",
            "type": "string"
          },
          "input_sha256": {
            "title": "Input Sha256",
            "type": "string"
          },
          "last_recovered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Recovered At"
          },
          "optimization_mode": {
            "$ref": "#/components/schemas/OptimizationMode"
          },
          "recovery_count": {
            "minimum": 0.0,
            "title": "Recovery Count",
            "type": "integer"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "run_type": {
            "$ref": "#/components/schemas/RunType"
          },
          "scenario_ids": {
            "items": {
              "type": "string"
            },
            "title": "Scenario Ids",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "v2g_enabled": {
            "title": "V2G Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "status",
          "run_type",
          "optimization_mode",
          "input_reference",
          "input_sha256",
          "v2g_enabled",
          "agent_backend",
          "scenario_ids",
          "requested_by",
          "created_at",
          "updated_at",
          "recovery_count"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "RunStatus": {
        "description": "Durable lifecycle states exposed by the demonstrator API.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "infeasible",
          "timed_out",
          "cancel_requested",
          "cancelled",
          "degraded"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "RunType": {
        "enum": [
          "day_ahead",
          "real_time"
        ],
        "title": "RunType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Human-approved decision support for electric-fleet charging and real-time replanning. This API does not control physical assets.",
    "title": "DepotFlux Industry Demonstrator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gridtwin/meta": {
      "get": {
        "operationId": "gridtwin_meta_api_v1_gridtwin_meta_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Operator-ID",
            "required": false,
            "schema": {
              "default": "demo-operator",
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Operator-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Gridtwin Meta Api V1 Gridtwin Meta Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Describe the synthetic GridTwin evidence experiment",
        "tags": [
          "gridtwin"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Operator-ID",
            "required": false,
            "schema": {
              "default": "demo-operator",
              "maxLength": 128,
              "minLength": 1,
              "title": "X-Operator-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a persisted optimization run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}